move classroom card component out of pages folder
